Point class moved into GrothendieckTopos
