+++ scala3/src/test/scala/math/cat/topos/logic/ConjunctionTest.scala
@@ -1,7 +1,7 @@
 package math.cat.topos.logic
 
 import math.cat.topos.CategoryOfDiagrams.DiagramArrow
-import math.cat.topos.{CategoryOfDiagrams, Fixtures, GrothendieckTopos, Point}
+import math.cat.topos.{CategoryOfDiagrams, Fixtures, GrothendieckTopos}
 import math.cat.{Category, SetFunction}
 import SetFunction.fun
 import org.specs2.execute.Result as MatchResult
